refactor: popconfirm types
