+++ projects/ion/src/lib/core/types/index.ts
@@ -16,6 +16,7 @@ export * from './input-counter';
 export * from './message';
 export * from './notification';
 export * from './pagination';
+export * from './popconfirm';
 export * from './status';
 export * from './tab-group';
 export * from './tooltip';
