+++ projects/ion/src/lib/core/types/popconfirm.ts
@@ -1,0 +1,9 @@
+import { StatusType } from './status';
+
+export interface PopConfirmProps {
+  ionPopConfirmTitle: string;
+  ionPopConfirmType?: StatusType;
+  ionPopConfirmDesc?: string;
+  ionConfirmText?: string;
+  ionCancelText?: string;
+}
